+++ android/lib/daemon-grpc/src/main/kotlin/net/mullvad/mullvadvpn/lib/daemon/grpc/mapper/ToDomain.kt
@@ -93,7 +93,7 @@ internal fun ManagementInterface.TunnelState.toDomain(): TunnelState =
                             location.toDomain()
                         } else null
                     },
-                featureIndicators = connected.featureIndicators.toDomain(),
+                featureIndicators = connecting.featureIndicators.toDomain(),
             )
         ManagementInterface.TunnelState.StateCase.CONNECTED ->
             TunnelState.Connected(
@@ -136,7 +136,7 @@ internal fun ManagementInterface.GeoIpLocation.toDomain(): GeoIpLocation =
         city = city,
         latitude = latitude,
         longitude = longitude,
-        hostname = hostname,
+        hostname = if (hasHostname()) hostname else null,
     )
 
 internal fun ManagementInterface.TunnelEndpoint.toDomain(): TunnelEndpoint =
@@ -605,7 +605,7 @@ internal fun ManagementInterface.SocksAuth.toDomain(): SocksAuth =
     SocksAuth(username = username, password = password)
 
 internal fun ManagementInterface.FeatureIndicators.toDomain(): List<FeatureIndicator> =
-    this.activeFeaturesList.map { it.toDomain() }
+    this.activeFeaturesList.map { it.toDomain() }.sorted()
 
 internal fun ManagementInterface.FeatureIndicator.toDomain() =
     when (this) {
@@ -627,5 +627,6 @@ internal fun ManagementInterface.FeatureIndicator.toDomain() =
         ManagementInterface.FeatureIndicator.MULTIHOP,
         ManagementInterface.FeatureIndicator.BRIDGE_MODE,
         ManagementInterface.FeatureIndicator.CUSTOM_MSS_FIX,
-        ManagementInterface.FeatureIndicator.UNRECOGNIZED -> error("Feature not supported")
+        ManagementInterface.FeatureIndicator.UNRECOGNIZED ->
+            error("Feature not supported ${this.name}")
     }